requirements.txt
python-binance==1.0.10
colorama==0.4.4
PyYAML==5.4.1
